Move name/name.go to name.go
